Created Escape Room Game Controller
Realised a use for my Media Player App. I have adapted it into a timer and background audio playing system for the Escape room I work at. Before we we had to use a couple of different systems which didn't work well and had a lot of room for human error. This system will easily allow the Games Master to control the room times and audio all-in-one panel!